Platforms Implemeted
Implemented moving platforms and elevators.
Elevators wait for a player to step on it before moving, which then waits a delay before moving to a target. If the goBackToStart bool is set to true in the inspector, then the elevator will go back to it's starting position automatically once the player has left the elevator.
Moving platforms move from Point A to Point B cyclically. You may set the pause timer to add a pause from when the platform gets to the point and it's following movement